create prop styles for individual and organization
